Enum sandwich::ProtoBasedErrorCode

pub enum ProtoBasedErrorCode {

    APIError(APIError),
    ConfigurationError(ConfigurationError),
    ProtobufError(ProtobufError),
    TLSConfigurationError(TLSConfigurationError),
    CertificateError(CertificateError),
    PrivateKeyError(PrivateKeyError),
    ASN1Error(ASN1Error),
    ALPNError(ALPNError),
    DataSourceError(DataSourceError),
    KEMError(KEMError),
    SystemError(SystemError),
    SocketError(SocketError),
    HandshakeError(HandshakeError),
    TunnelError(TunnelError),

}

An error code. An error code holds one of the error described in errors.proto and listed in ErrorKind.

Variants

APIError(APIError)

API errors. The following errors can occur during a call to the Context API.

ConfigurationError(ConfigurationError)

Errors regarding configurations.

ProtobufError(ProtobufError)

Errors regarding protobuf.

TLSConfigurationError(TLSConfigurationError)

Errors regarding TLS configurations.

CertificateError(CertificateError)

Certificate errors.

PrivateKeyError(PrivateKeyError)

Private key errors.

ASN1Error(ASN1Error)

ASN.1 errors.

ALPNError(ALPNError)

ALPN errors.

DataSourceError(DataSourceError)

DataSource errors.

KEMError(KEMError)

KEM errors.

SystemError(SystemError)

System errors.

SocketError(SocketError)

Socket errors. These errors are used in io/socket.

HandshakeError(HandshakeError)

TunnelError(TunnelError)

Tunnel error.