Trait sandwich::io::listener::Listener

pub trait Listener: Send {
    // Required methods
    fn listen(&mut self) -> Result<(), Error>;
    fn accept(&mut self) -> Result<Box<dyn IO>, Error>;
    fn close(&mut self) -> Result<(), Error>;

    // Provided method
    fn destructor(&mut self) -> Result<(), Error> { ... }
}

A listener for connections.

Required Methods

fn listen(&mut self) -> Result<(), Error>

Similar to the linux system call listen. Sets up the listener to accept incoming connections.

fn accept(&mut self) -> Result<Box<dyn IO>, Error>

Similar to the linux system call accept. Accepts incoming connection requests based on parameters set by self.listen().

fn close(&mut self) -> Result<(), Error>

Closes the listener and rejects all future connection requests.

Provided Methods

fn destructor(&mut self) -> Result<(), Error>

A custom destructor to be called when memory is freed.